Added s=150 to the avatars.
Assumed the avatar URL returned by the GitHub API will always have
a querystring. It currently does, so I've just appended `s=150` onto the
end. When I tested it too, the last value for `s` wins, so if you
already have `s=80` already in the URL and we append `s=150` we'll get
150x150 pixel images.

Removed _Contributors.cshtml, as all the HTML for this has been moved
into the Mustache script template that sits in Views/Home/Index.cshtml.

I've added an octocat background image to the #contributors element so
that something is displayed while we're loading the contributors via the
API.

I've also updated the CSS for the contributors list, so that it displays
the same way as the current homepage is displayed.